Add resident demerit records and demerit command

+++ src/main/java/seedu/address/logic/parser/AddressBookParser.java
@@ -12,6 +12,7 @@ import seedu.address.logic.commands.AddCommand;
 import seedu.address.logic.commands.ClearCommand;
 import seedu.address.logic.commands.Command;
 import seedu.address.logic.commands.DeleteCommand;
+import seedu.address.logic.commands.DemeritCommand;
 import seedu.address.logic.commands.DemeritListCommand;
 import seedu.address.logic.commands.EditCommand;
 import seedu.address.logic.commands.ExitCommand;
@@ -56,41 +57,31 @@ public class AddressBookParser {
         // Lower level log messages are used sparingly to minimize noise in the code.
         logger.fine("Command word: " + commandWord + "; Arguments: " + arguments);
 
-        switch (commandWord) {
-        case AddCommand.COMMAND_WORD:
+        if (commandWord.equals(AddCommand.COMMAND_WORD)) {
             return new AddCommandParser().parse(arguments);
-
-        case EditCommand.COMMAND_WORD:
+        } else if (commandWord.equals(EditCommand.COMMAND_WORD)) {
             return new EditCommandParser().parse(arguments);
-
-        case DeleteCommand.COMMAND_WORD:
+        } else if (commandWord.equals(DeleteCommand.COMMAND_WORD)) {
             return new DeleteCommandParser().parse(arguments);
-
-        case ClearCommand.COMMAND_WORD:
+        } else if (commandWord.equals(ClearCommand.COMMAND_WORD)) {
             return new ClearCommand();
-
-        case FindCommand.COMMAND_WORD:
+        } else if (commandWord.equals(FindCommand.COMMAND_WORD)) {
             return new FindCommandParser().parse(arguments);
-
-        case ListCommand.COMMAND_WORD:
+        } else if (commandWord.equals(ListCommand.COMMAND_WORD)) {
             return new ListCommand();
-
-        case ExitCommand.COMMAND_WORD:
+        } else if (commandWord.equals(ExitCommand.COMMAND_WORD)) {
             return new ExitCommand();
-
-        case HelpCommand.COMMAND_WORD:
+        } else if (commandWord.equals(HelpCommand.COMMAND_WORD)) {
             return new HelpCommand();
-
-        case TagCommand.COMMAND_WORD:
+        } else if (commandWord.equals(TagCommand.COMMAND_WORD)) {
             return new TagCommandParser().parse(arguments);
-
-        case RemarkCommand.COMMAND_WORD:
+        } else if (commandWord.equals(RemarkCommand.COMMAND_WORD)) {
             return new RemarkCommandParser().parse(arguments);
-
-        case DemeritListCommand.COMMAND_WORD:
+        } else if (commandWord.equals(DemeritListCommand.COMMAND_WORD)) {
             return new DemeritListCommand();
-
-        default:
+        } else if (commandWord.equals(DemeritCommand.COMMAND_WORD)) {
+            return new DemeritCommandParser().parse(arguments);
+        } else {
             logger.finer("This user input caused a ParseException: " + userInput);
             throw new ParseException(MESSAGE_UNKNOWN_COMMAND);
         }
